refactor: 페이지네이션 시, 파라미터로 size와 page를 입력받기
